// apps/web/src/contracts/generated/sheet/MapContent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RasterMap } from "./RasterMap";
import type { VectorMap } from "./VectorMap";

/**
 * A map frame's content, drawn in its frame, cut to it.
 */
export type MapContent = { "type": "vector" } & VectorMap | { "type": "raster" } & RasterMap;